[4] Rewrite the results-parser in C++ to speed it up tremendously

+++ parseresults.sh
@@ -11,6 +11,11 @@ then
   echo "Cannot use sqlite3 command. Please check if SQLite3 is installed on this system and you have permission to use it." >&2
   exit 1
 fi
+if [ ! -x ./parseresults ]
+then
+  echo "Cannot use parseresults binary. See ./cpp directory for more information. You dit run install.sh before using this script, right?" >&2
+  exit 1
+fi
 if [ -f results.db ]
 then
   rm results.db
@@ -26,24 +31,8 @@ $SEDCMD s/%27/g c*.txt
 echo "Creating results database in results.db"
 $SQLITECMD results.db < ddl.sql
 
-for resultfile in c*.txt
+for counter in $(seq $START $INCREMENT $END)
 do
-  echo "Processing data in $resultfile, this may take a while!"
-  echo "BEGIN TRANSACTION; " > queryfile.sql
-  cValue="$(grep -o -E '[0-9]{2,}' <<< $resultfile)"
-  while read line; do
-    if grep -q  '^HTTP/' <<< $line
-    then
-      PROTO=$(echo $line | cut -f1 -d,)
-      STATUS=$(echo $line | cut -f2 -d,)
-      VOLUME=$(echo $line | cut -f4 -d,)
-      RT=$(echo $line | cut -f3 -d,)
-      URL=$(echo $line | cut -f5 -d,)
-      DATETIME=$(echo $line | cut -f7 -d,)
-      echo "INSERT INTO main.results (PROTO,STATUS,VOLUME,RT,URL,DATETIME,C) VALUES ('"$PROTO"',$STATUS,$VOLUME,$RT,'"$URL"','"$DATETIME"',$cValue);" >>queryfile.sql
-    fi
-  done < $resultfile
-  echo "COMMIT TRANSACTION;" >>queryfile.sql
-  $SQLITECMD results.db < queryfile.sql
+  echo "Working on C-level $counter." >&2
+  ./parseresults $counter
 done
-rm queryfile.sql
